+++ components/shys_m5_dial/ha_device_mode_mediaplayer_play.h
@@ -17,7 +17,7 @@ namespace esphome
                 
                 unsigned long lastRefresh = 0;
 
-                void sendValueToHomeAssistant(int value) override {
+                void sendValueToHomeAssistant(float value) override {
                     haApi.setMediaPlayerVolume(this->device.getEntityId(), value);
                 }
 
@@ -261,7 +261,7 @@ namespace esphome
 
                 }
 
-                bool onTouch(M5DialDisplay& display, uint16_t x, uint16_t y) override {
+                bool onTouch(M5DialDisplay& display, float x, float y) override {
                     ESP_LOGI("TOUCH", "Touch %i / %i for %s", x, y, this->device.getEntityId().c_str());
 
                     LovyanGFX* gfx = display.getGfx();
